requirements.txt
gradio
tensorflow
opencv-python
Pillow
numpy